Rename utils module to transformer module
